refactor: split incremental runner + move remote workflow body
DiagramGenerator.generate_analysis_incremental is now a thin delegate
to a new IncrementalRunner class (diagram_analysis/incremental_runner.py)
that owns the structural delta + LLM tracing pipeline. The runner takes
the generator as a collaborator so it can reach the helpers
(_normalize_repo_path, _build_file_coverage_summary, etc.) without
duplicating them. Eight imports are now dead in diagram_generator.py
(IncrementalUpdater, run_trace, etc.) — dropped.

Also: codeboarding_cli/commands/full_analysis.py's _process_one_remote
(50-line workflow with monitoring + render_docs + artifact copy) moved
into a new codeboarding_workflows/remote.py:process_one_remote. The
command file is back to argparse plumbing + a single workflow call,
matching how full_remote and partial were always organized.

Why: addresses the architectural-review claims that
(a) generate_analysis_incremental was a parallel pipeline hiding inside
DiagramGenerator and (b) command handlers were doing real workflow work
inline instead of calling into codeboarding_workflows.

Tests: 1557 pass (same 2 pre-existing failures unrelated to this work).
